requirements/development.txt
-r base.txt

pre-commit==3.7.1
ruff==0.5.4
black==24.4.2
django-debug-toolbar==4.4.6
django-rest-framework==0.1.0
drf-yasg==1.21.8